book_collection_api/books/requirement.txt
Django>=3.0,<4.0
djangorestframework>=3.11,<4.0
drf-yasg>=1.20,<2.0
